pyproject.toml
[project]
name = "flightplotting"
version="v0.2.13"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Thomas David", email = "thomasdavid0@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "flightdata>=0.3.0",
    "pfc-geometry",
    "pfcschemas",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/plotting"]

[dependency-groups]
dev = ["pytest>=8.3.3"]
